Media create method. Create s3 CRUD service.
