+++ src/middleware/s3-file-upload.ts
@@ -7,7 +7,7 @@ const env = process.env.NODE_ENV;
 const envConfig = config[env];
 const { region, accessKeyId, secretAccessKey, bucket } = envConfig.s3;
 
-const s3 = new S3( { accessKeyId, secretAccessKey, region } );
+export const s3 = new S3( { accessKeyId, secretAccessKey, region } );
 
 export const uploadToS3 = (): multer.Instance => {
 
@@ -25,13 +25,3 @@ export const uploadToS3 = (): multer.Instance => {
   });
 
 };
-
-export const getPreSignedUrl = (key: string, expires: number): Promise<string> => {
-  return new Promise ((res, rej) => {
-    const params = { Bucket: bucket, Key: key, Expires: expires };
-    s3.getSignedUrl("getObject", params, (err, url) => {
-      if (err) rej(err);
-      res(url);
-    });
-  });
-}
